eithoneydew: Merge from HEAD. BUGZID:12485
git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/eithoneydew@9965 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ uniconf/uniinigen.cc
@@ -5,11 +5,11 @@
  * A generator for .ini files.
  */
 #include "uniinigen.h"
-#include "unitempgen.h"
-#include "wvtclstring.h"
 #include "strutils.h"
+#include "unitempgen.h"
 #include "wvfile.h"
 #include "wvmoniker.h"
+#include "wvtclstring.h"
 #include <ctype.h>
 #include "wvlinkerhack.h"
 
@@ -39,6 +39,7 @@ UniIniGen::UniIniGen(WvStringParm _filename, int _create_mode)
     //log(WvLog::Debug1, "Using IniFile \"%s\"\n", filename);
     // consider the generator dirty until it is first refreshed
     dirty = true;
+    memset(&old_st, 0, sizeof(old_st));
 }
 
 void UniIniGen::set(const UniConfKey &key, WvStringParm value)
@@ -58,7 +59,7 @@ bool UniIniGen::refresh()
 {
     WvFile file(filename, O_RDONLY);
 
-    #ifndef _WIN32
+#ifndef _WIN32
     struct stat statbuf;
     if (file.isok() && fstat(file.getrfd(), &statbuf) == -1)
     {
@@ -72,7 +73,19 @@ bool UniIniGen::refresh()
 	file.close();
 	file.seterr(EAGAIN);
     }
-    #endif
+    
+    if (file.isok() // guarantes statbuf is valid from above
+	&& statbuf.st_ctime == old_st.st_ctime
+	&& statbuf.st_dev == old_st.st_dev
+	&& statbuf.st_ino == old_st.st_ino
+	&& statbuf.st_blocks == old_st.st_blocks
+	&& statbuf.st_size == old_st.st_size)
+    {
+	log(WvLog::Debug, "refresh: file hasn't changed; do nothing.\n");
+	return true;
+    }
+    memcpy(&old_st, &statbuf, sizeof(statbuf));
+#endif
 
     if (!file.isok())
     {
@@ -207,6 +220,7 @@ bool UniIniGen::refresh()
 
     WVRELEASE(newgen);
 
+    UniTempGen::refresh();
     return true;
 }
 
@@ -247,32 +261,43 @@ bool UniIniGen::refreshcomparator(const UniConfValueTree *a,
 
 
 #ifndef _WIN32
-bool UniIniGen::commit_atomic(WvString real_filename)
+bool UniIniGen::commit_atomic(WvStringParm real_filename)
 {
+    struct stat statbuf;
+
+    if (lstat(real_filename, &statbuf) == -1)
+    {
+	if (errno != ENOENT)
+	    return false;
+    }
+    else
+	if (!S_ISREG(statbuf.st_mode))
+	    return false;
+
     WvString tmp_filename("%s.tmp%s", real_filename, getpid());
     WvFile file(tmp_filename, O_WRONLY|O_TRUNC|O_CREAT, 0000);
-    struct stat statbuf;
-    
-    if (file.geterr()
-	|| lstat(real_filename, &statbuf) == -1
-	|| !S_ISREG(statbuf.st_mode))
+
+    if (file.geterr())
     {
-        log(WvLog::Warning, "Can't write '%s': %s\n",
-                filename, strerror(errno));
+	log(WvLog::Warning, "Can't write '%s': %s\n",
+	    filename, strerror(errno));
 	unlink(tmp_filename);
-        file.close();
-        return false;
+	file.close();
+	return false;
     }
-    
+
     save(file, *root); // write the changes out to our temp file
-    
+
+    mode_t theumask = umask(0);
+    umask(theumask);
+    fchmod(file.getwfd(), create_mode & ~theumask);
+
     file.close();
-    chmod(tmp_filename, create_mode);
-    if (rename(tmp_filename, real_filename) == -1
-            || file.geterr())
+
+    if (file.geterr() || rename(tmp_filename, real_filename) == -1)
     {
         log(WvLog::Warning, "Can't write '%s': %s\n",
-                filename, strerror(errno));
+	    filename, strerror(errno));
         unlink(tmp_filename);
 	return false;
     }
@@ -284,11 +309,14 @@ bool UniIniGen::commit_atomic(WvString real_filename)
 
 void UniIniGen::commit()
 {
-    if (!dirty) return;
+    if (!dirty)
+	return;
+
+    UniTempGen::commit();
 
 #ifdef _WIN32
-    // Windows doesn't support all that fancy stuff, just open the file
-    //   and be done with it
+    // Windows doesn't support all that fancy stuff, just open the
+    // file and be done with it
     WvFile file(filename, O_WRONLY|O_TRUNC|O_CREAT, create_mode);
     save(file, *root); // write the changes out to our file
     file.close();
@@ -299,17 +327,14 @@ void UniIniGen::commit()
 	return;
     }
 #else
-    char resolved_path[PATH_MAX];
-    WvFile file(filename, O_WRONLY|O_TRUNC|O_CREAT, 0000);
     WvString real_filename(filename);
+    char resolved_path[PATH_MAX];
 
     if (realpath(filename, resolved_path) != NULL)
 	real_filename = resolved_path;
 
-    // first try to overwrite the file atomically
     if (!commit_atomic(real_filename))
     {
-        // if not, overwrite it in place
         WvFile file(real_filename, O_WRONLY|O_TRUNC|O_CREAT, create_mode);
         struct stat statbuf;
 
@@ -321,7 +346,8 @@ void UniIniGen::commit()
         }
 
         fchmod(file.getwfd(), (statbuf.st_mode & 07777) | S_ISVTX);
-        save(file, *root); // write the changes out to our file
+
+        save(file, *root);
     
         if (!file.geterr())
         {
@@ -333,9 +359,7 @@ void UniIniGen::commit()
 	else
 	    log(WvLog::Warning, "Error writing '%s' ('%s'): %s\n",
 		filename, real_filename, file.errstr());
-
     }
-    file.close();
 #endif
 
     dirty = false;
